MSP-2788: moved pod template from common to prometheus package
